refactor(e2e): designer/e2e/mcp/ の WS ヘルパーを _helpers.ts に共通化 (#590)
PR #589 Sonnet レビュー Should-fix #2 対応。
isMcpRunning / sendBrowserRequest が mcp-tools.spec.ts と
bulk-load.spec.ts で完全重複していたため、
designer/e2e/mcp/_helpers.ts に抽出して両 spec から import に置換する。

挙動変更なし、リファクタのみ。

Closes #590
